Enable warnings for uninitialized variables
This extends the code for handling uninitialized output parameters.
Still needs to handle generic templates and assignment of uninitialized
values more carefully.

The file containing the relevant code are now in

    source/slang/slang-ir-use-uninitialized-values.cpp

rather than the previous

    source/slang/slang-ir-use-uninitialized-out-param.h

and the top-level function is now checkForUsingUinitializedValues.

Additionally a rudimentary test shader has been added for this case.
