+++ tests/diagnostics/uninitialized-variables.slang
@@ -1,0 +1,12 @@
+int ok(int x)
+{
+    int k;
+    return k;
+}
+
+int error()
+{
+    int x;
+    int k;
+    x = k;
+}
